liquida2018/requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
django-bootstrap4==0.0.6
django-currentuser==0.2.2
django-filter==1.1.0
django-localflavor==2.0
django-localflavor-br==1.0
django-widget-tweaks==1.4.2
idna==2.6
oauthlib==2.1.0
Pillow==5.1.0
PyJWT==1.6.4
python-social-auth==0.2.12
python3-openid==3.1.0
pytz==2018.4
qrcode==5.1
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.22
